use fallbackComponent instead
