--- addons/compat_rhs_afrf3_usf3/$PBOPREFIX$
@@ -1,1 +1,0 @@
-z\ace\addons\compat_rhs_afrf3_usf3
