--- addons/compat_rhs_afrf3_usf3/XEH_preInit.sqf
@@ -1,10 +1,0 @@
-#include "script_component.hpp"
-
-ADDON = false;
-
-PREP_RECOMPILE_START;
-#include "XEH_PREP.hpp"
-PREP_RECOMPILE_END;
-
-
-ADDON = true;
