--- addons/compat_rhs_afrf3_usf3/XEH_preStart.sqf
@@ -1,3 +1,0 @@
-#include "script_component.hpp"
-
-#include "XEH_PREP.hpp"
