--- addons/compat_rhs_afrf3_usf3/compat_rhs_afrf3_usf3_wardrobe/CfgWardrobe.hpp
@@ -1,6 +1,0 @@
-class ace_wardrobe {
-    #include "\z\ace\addons\wardrobe\BaseClasses_Import.hpp"
-
-    #include "CfgWardrobe_Uniforms.hpp"
-    #include "CfgWardrobe_Headgear.hpp"
-};
