--- addons/compat_rhs_afrf3_usf3/compat_rhs_afrf3_usf3_wardrobe/CfgWardrobe_Uniforms.hpp
@@ -1,8 +1,0 @@
-UNIFORM_GLOVES(rhs_uniform_6sh122_gloves_v1,rhs_uniform_6sh122_v1);
-UNIFORM_GLOVES(rhs_uniform_6sh122_gloves_v2,rhs_uniform_6sh122_v2);
-
-UNIFORM_GLOVES(rhs_uniform_gorka_r_g_gloves,rhs_uniform_gorka_r_g);
-UNIFORM_GLOVES(rhs_uniform_gorka_r_y_gloves,rhs_uniform_gorka_r_y);
-
-UNIFORM_GLOVES(rhs_uniform_vkpo_gloves,rhs_uniform_vkpo);
-UNIFORM_GLOVES(rhs_uniform_vkpo_gloves_alt,rhs_uniform_vkpo_alt);
