--- addons/compat_rhs_afrf3_usf3/compat_rhs_afrf3_usf3_wardrobe/CfgWeapons.hpp
@@ -1,2 +1,0 @@
-class CfgWeapons {
-};
