--- addons/compat_rhs_afrf3_usf3/compat_rhs_afrf3_usf3_wardrobe/config.cpp
@@ -1,27 +1,0 @@
-#include "script_component.hpp"
-
-class CfgPatches {
-    class SUBADDON {
-        name = COMPONENT_NAME;
-        units[] = {};
-        weapons[] = {};
-        requiredVersion = REQUIRED_VERSION;
-        requiredAddons[] = {
-            "ace_wardrobe",
-            "rhs_main_loadorder",
-            "rhsusf_main_loadorder"
-        };
-        skipWhenMissingDependencies = 1;
-        author = ECSTRING(common,ACETeam);
-        authors[] = {"OverlordZorn [CVO]", "Andx [TTT]"};
-        url = ECSTRING(main,URL);
-        VERSION_CONFIG;
-        addonRootClass = QUOTE(ADDON);
-    };
-};
-
-// For Consistency: Adjust Mass/containerSize
-#include "CfgWeapons.hpp"
-
-// ACE Wardrobe
-#include "CfgWardrobe.hpp"
