--- addons/compat_rhs_afrf3_usf3/compat_rhs_afrf3_usf3_wardrobe/script_component.hpp
@@ -1,5 +1,0 @@
-#define SUBCOMPONENT wardrobe
-#define SUBCOMPONENT_BEAUTIFIED Wardrobe
-#include "..\script_component.hpp"
-
-#include "\z\ace\addons\wardrobe\script_macros_wardrobe.hpp"
